pipelines/requirements.txt

boto3==1.35.82
botocore
buildingregulariser
geopandas
geoserver-rest==2.10.0
humanize
ibm-cos-sdk
Jinja2
joblib
matplotlib
numpy==2.3.2
opentelemetry-distro
opentelemetry-exporter-otlp
pandas
pg8000
python-dotenv
rasterio
requests
rio-cogeo
rioxarray
scikit-learn>=1.3.0
sentinelhub
shapely
sqlalchemy
tenacity
terrakit==0.1.3
tqdm
wget
xarray



  

